pyproject.toml
[project]
name = "earnings-crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Vutala", email = "wesly.vutala@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "finnhub-python>=2.4.28",
    "google-genai>=1.73.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "twelvedata>=1.3.0",
]

[project.scripts]
earnings-crew = "earnings_crew:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
